Create HashMap class and add set method
The set method calculates the code of the Node to be inserted using
the hash method with the key as an argument. The remainder of the code
divided by the amount of buckets is counted as the future index of the
object.

If a LinkedList was not yet created at that index, the method creates one
and uses append to make the new Node.

If a LinedList was already created, it simply appends the new Node to it
and updates the next_node value of the nodes.