Add #place_color method
The method asks for player input, formats it into an array of integers
and assigns it to the corresponding row/col pair depending on the round.

Before doing that, it makes sure that the space is empty. If it isn't, it
puts an error message and asks for different input.